requirements.txt
flask
requests
faiss-cpu
python-dotenv